claude-helpers/claude-flow/pyproject.toml
[project]
name = "claude-workflow-kanban"
version = "0.1.0"
description = "Kanban board backend for Claude Code workflow management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.18",
    "pydantic-ai>=0.0.30",
    "python-dotenv>=1.0.0",
    "pywebview>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # Depends() in defaults is standard FastAPI pattern
    "ARG001", # lifespan requires app arg per FastAPI spec
]

[tool.ruff.lint.isort]
known-first-party = ["kanban"]

[tool.hatch.build.targets.wheel]
packages = ["src/kanban"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
